# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample hotel API",
        "title": "Hotel API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/create_hotel": {
            "post": {
                "description": "Create a new hotel with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new hotel",
                "operationId": "create-hotel",
                "parameters": [
                    {
                        "description": "Hotel Payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Hotel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created hotel with ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
        "model.Hotel": {
            "type": "object",
            "required": [
                "description",
                "location_id",
                "name",
                "number",
                "worker_id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "integer"
                }
            }
        }
    }
}